algoshield/algoshield/requirements.txt
algopy==0.7.2
asgiref==3.8.1
cffi==1.17.1
colorama==0.4.6
Django==5.1.2
djangorestframework==3.15.2
docstring-parser==0.14.1
executing==2.0.1
iniconfig==2.0.0
msgpack==1.1.0
numpy==2.1.2
packaging==24.1
pluggy==1.5.0
py-algorand-sdk==2.6.1
pycparser==2.22
pycryptodomex==3.21.0
PyNaCl==1.5.0
pyteal==0.26.1
pytest==8.3.3
scipy==1.14.1
semantic-version==2.10.0
sqlparse==0.5.1
tabulate==0.9.0
tzdata==2024.2
python-dotenv==1.0.1
django-cors-headers==4.6.0
axios==0.4.0
gunicorn==23.0.0 
whitenoise==6.8.1
beaker-pyteal==1.1.1

